# docs/specs/openapi/nauthilus-director.yaml
openapi: 3.0.3
info:
  title: nauthilus-director control API
  version: v1
  description: >
    Administrative control API for nauthilus-director. The v1 surface is
    OpenAPI-first; generated server and client code stay at the REST boundary.
...
servers:
  - url: http://127.0.0.1:9090
    description: Default local control listener
...
paths:
...
  /api/v1/route/lookup:
    post:
      operationId: lookupRoute
      tags:
        - routing
      summary: Explain director-only routing for a known identity key.
      description: >
        This diagnostic endpoint is side-effect-free. It must not authenticate
        credentials, call Nauthilus, create sessions, refresh leases or mutate
        Redis state.
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/RouteLookupRequest"
      responses:
        "200":
          description: Route lookup explanation.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/RouteLookupResponse"
        "400":
          $ref: "#/components/responses/BadRequest"
        "501":
          $ref: "#/components/responses/NotImplemented"
        default:
          $ref: "#/components/responses/Error"
...
components:
...
  responses:
    BadRequest:
      description: The request was rejected before domain processing.
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/ErrorResponse"
    Error:
      description: Structured error response.
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/ErrorResponse"
    NotImplemented:
      description: The route is reserved but the domain behavior is not implemented.
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/ErrorResponse"
  schemas:
...
    ErrorResponse:
      type: object
      additionalProperties: false
      required:
        - status
        - code
        - message
      properties:
        status:
          type: integer
          minimum: 100
          maximum: 599
        code:
          type: string
          minLength: 1
        message:
          type: string
          minLength: 1
        operation:
          type: string
          nullable: true
...
    RouteLookupRequest:
      type: object
      additionalProperties: false
      required:
        - protocol
        - user_key
      properties:
        protocol:
          type: string
          minLength: 1
        user_key:
          type: string
          minLength: 1
        tenant:
          type: string
        listener:
          type: string
        service_name:
          type: string
        backend_pool:
          type: string
        client_ip:
          type: string
        attributes:
          type: object
          additionalProperties:
            type: array
            items:
              type: string
    RouteLookupResponse:
      type: object
      additionalProperties: false
      required:
        - selected_backend
        - reason
        - healthy
        - maintenance
        - shard_tag
      properties:
        selected_backend:
          type: string
        reason:
          type: string
        healthy:
          type: boolean
        maintenance:
          type: boolean
        shard_tag:
          type: string
        routing_generation:
          type: string
...